Move SchemaAddFieldModal to its own folder
+ misc test cleanup - remove unnecessary jest.spyOn in individual tests, it's already in the beforeEach
