+++ x-pack/plugins/enterprise_search/public/applications/shared/schema/index.ts
@@ -5,6 +5,6 @@
  * 2.0.
  */
 
-export { SchemaAddFieldModal } from './schema_add_field_modal';
+export { SchemaAddFieldModal } from './add_field_modal';
 export { SchemaFieldTypeSelect } from './field_type_select';
 export { SchemaErrorsCallout } from './errors_callout';
